Cleanup metadata model. Old model is now DeprecatedBookMetadata. MultiLangMetadata is now BookMetadata.
