+++ rosa-archive/rosa-archive-model/src/main/java/rosa/archive/model/ArchiveItemType.java
@@ -5,8 +5,8 @@ public enum ArchiveItemType {
     CROPPED_IMAGE_LIST("images.crop.csv", "csv"),
     IMAGE_LIST("images.csv", "csv"),
     CROPPING("crop", "txt"),
-    DESCRIPTION("description_", "xml"),
-    DESCRIPTION_MULTILANG("description.xml", "xml"),
+    DEPRECATED_DESCRIPTION("description_", "xml"),
+    METADATA("description.xml", "xml"),
     FILE_MAP("filemap", "csv"),
     SHA1SUM("", "SHA1SUM"),
     NARRATIVE_TAGGING("nartag", "csv"),
