+++ rosa-archive/rosa-archive-model/src/main/java/rosa/archive/model/BiblioData.java
@@ -1,10 +1,10 @@
-package rosa.archive.model.meta;
+package rosa.archive.model;
 
 import java.io.Serializable;
 import java.util.Arrays;
 
 /**
- *
+ * Bilbliographic data in a given language.
  */
 public class BiblioData implements Serializable {
     private static final long serialVersionUID = 1L;
